requirements.txt
django~=3.2.5
djangorestframework~=3.12.4
gunicorn
django-heroku
dj-config-url~=0.1.1
dj_database_url
whitenoise
boto3~=1.20.31
botocore~=1.23.31
djangorestframework-api-key~=2.1.0
docopt~=0.6.2
webvtt-py~=0.4.6
django-cors-headers~=3.11.0
pdf2image~=1.16.0